requirements.txt
streamlit
sentence-transformers
chromadb
llama-cpp-python
arize-phoenix 